notice-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
motor==3.6.0
beautifulsoup4==4.12.3
prometheus-fastapi-instrumentator==7.0.0
